Moved to package 'bukkit'
