+++ gradle.properties
@@ -16,5 +16,5 @@
 # along with BukkitMimic.  If not, see <http://www.gnu.org/licenses/>.
 #
 
-pluginMain = ru.endlesscode.mimic.BukkitMimic
+pluginMain = ru.endlesscode.mimic.bukkit.BukkitMimic
 url = https://github.com/EndlessCodeGroup/BukkitMimic
