+++ src/main/java/ru/endlesscode/mimic/bukkit/system/PermissionsClassSystem.java
@@ -16,7 +16,7 @@
  * along with BukkitMimic.  If not, see <http://www.gnu.org/licenses/>.
  */
 
-package ru.endlesscode.mimic.system;
+package ru.endlesscode.mimic.bukkit.system;
 
 import com.google.common.annotations.VisibleForTesting;
 import org.bukkit.entity.Player;
