+++ src/main/java/ru/endlesscode/mimic/bukkit/system/VanillaConverter.java
@@ -16,7 +16,7 @@
  * along with BukkitMimic.  If not, see <http://www.gnu.org/licenses/>.
  */
 
-package ru.endlesscode.mimic.system;
+package ru.endlesscode.mimic.bukkit.system;
 
 import org.jetbrains.annotations.NotNull;
 import ru.endlesscode.mimic.api.system.ExpLevelConverter;
