+++ src/test/java/ru/endlesscode/mimic/bukkit/system/TestSystem.java
@@ -16,7 +16,7 @@
  * along with BukkitMimic.  If not, see <http://www.gnu.org/licenses/>.
  */
 
-package ru.endlesscode.mimic.system;
+package ru.endlesscode.mimic.bukkit.system;
 
 import ru.endlesscode.mimic.api.system.PlayerSystem;
